Move applications_api to namespaced api file
